Support AMP GUUI endpoint is explicitly requested
Also refactor/tidyup some of the picker code as felt it could be
simpler and more readable.
