+++ article/app/AppLoader.scala
@@ -1,28 +1,28 @@
-import http.{CommonFilters, CorsHttpErrorHandler}
+import _root_.commercial.targeting.TargetingLifecycle
+import akka.actor.ActorSystem
 import app.{FrontendApplicationLoader, FrontendComponents}
 import assets.DiscussionExternalAssetsLifecycle
 import com.softwaremill.macwire._
-import common._
 import common.Logback.{LogbackOperationsPool, LogstashLifecycle}
-import _root_.commercial.targeting.TargetingLifecycle
-import akka.actor.ActorSystem
+import common._
 import common.commercial.OrielCacheLifecycle
 import common.dfp.DfpAgentLifecycle
-import conf.switches.SwitchboardLifecycle
 import conf.CachedHealthCheckLifeCycle
+import conf.switches.SwitchboardLifecycle
 import contentapi.{CapiHttpClient, ContentApiClient, HttpClient}
 import controllers.{ArticleControllers, HealthCheck}
 import dev.{DevAssetsController, DevParametersHttpRequestHandler}
+import http.{CommonFilters, CorsHttpErrorHandler}
 import model.ApplicationIdentity
-import services.ophan.SurgingContentAgentLifecycle
 import play.api.ApplicationLoader.Context
 import play.api.BuiltInComponentsFromContext
 import play.api.http.{HttpErrorHandler, HttpRequestHandler}
 import play.api.mvc.EssentialFilter
 import play.api.routing.Router
-import services.{NewspaperBooksAndSectionsAutoRefresh, OphanApi, SkimLinksCacheLifeCycle}
 import router.Routes
 import services.dotcomponents.RenderingTierPicker
+import services.ophan.SurgingContentAgentLifecycle
+import services.{NewspaperBooksAndSectionsAutoRefresh, OphanApi, SkimLinksCacheLifeCycle}
 
 class AppLoader extends FrontendApplicationLoader {
   override def buildComponents(context: Context): FrontendComponents = new BuiltInComponentsFromContext(context) with AppComponents
@@ -38,7 +38,7 @@ trait AppComponents extends FrontendComponents with ArticleControllers {
   lazy val devAssetsController = wire[DevAssetsController]
   lazy val logbackOperationsPool = wire[LogbackOperationsPool]
 
-  lazy val remoteRender = wire[renderers.RemoteRender]
+  lazy val remoteRender = wire[renderers.RemoteRenderer]
   lazy val renderingTierPicker = wire[RenderingTierPicker]
 
   override lazy val lifecycleComponents = List(
