+++ article/app/services/dotcomponents/RenderType.scala
@@ -2,4 +2,5 @@ package services.dotcomponents
 
 sealed trait RenderType
 case object RemoteRender extends RenderType
+case object RemoteRenderAMP extends RenderType
 case object LocalRender extends RenderType
