--- article/app/services/dotcomponents/pickers/RenderTierPickerStrategy.scala
@@ -1,9 +1,0 @@
-package services.dotcomponents.pickers
-
-import model.PageWithStoryPackage
-import play.api.mvc.RequestHeader
-import services.dotcomponents.RenderType
-
-trait RenderTierPickerStrategy {
-  def getRenderTierFor(page: PageWithStoryPackage, request: RequestHeader): (List[(String, Boolean)], RenderType)
-}
