+++ common/app/common/configuration.scala
@@ -148,6 +148,7 @@ class GuardianConfiguration extends Logging {
 
   object rendering {
     lazy val renderingEndpoint = configuration.getMandatoryStringProperty("rendering.endpoint")
+    lazy val AMPArticleEndpoint = configuration.getMandatoryStringProperty("rendering.AMPArticleEndpoint")
   }
 
   object weather {
